Update patch set 2
Patch Set 2:

Yes, we should fix that.

Could you cat the value when that happens?

IIRC the defaults are 5, 10 and 10 (dark, dim and minimum), so try setting those manually and see if it flickers.
Then increment the value one by one, and adjust the above 3 accordingly. You shouldn't need to change anything else.

Patch-set: 2