Update patch set 5
Patch Set 5: -Code-Review

I really don't know how do you want to do it as my knowledge is not this big so I will just let this sit here for the values because I noticed that values in msm8974-common are too low (there is 5 and 10 again). I hope that this will be done as soon as possible because this is the only thing that keeps me from using CM as my daily driver.

Patch-set: 5